Feat: Add popup globally withou index
