+++ src/index.js
@@ -1,14 +1,14 @@
 import './style.css';
 import displayUI from './modules/displayUI.js';
-import createPopup from './modules/popup.js';
+// import createPopup from './modules/popup.js';
 
 displayUI();
 
-window.addEventListener('load', () => {
-  const commentBtn = document.querySelectorAll('.btnClass');
-  commentBtn.forEach((element, index) => {
-    element.addEventListener('click', () => {
-      createPopup(index);
-    });
-  });
-});
+// window.addEventListener('load', () => {
+//   const commentBtn = document.querySelectorAll('.btnClass');
+//   commentBtn.forEach((element, index) => {
+//     element.addEventListener('click', () => {
+//       createPopup(index);
+//     });
+//   });
+// });
